requirements.txt
mcp[cli]>=1.26.0
httpx>=0.27.0
uvicorn[standard]>=0.30.0
fastapi>=0.111.0
pytest>=8.0.0
python-dotenv>=1.0.0
pytest-asyncio>=0.23.0
